# REF/PERF: MultiIndex.get_indexer with method