--- a/pandas/core/indexes/base.py
+++ b/pandas/core/indexes/base.py
@@ -4023,17 +4023,12 @@ class Index(IndexOpsMixin, PandasObject):
         if self._is_multi:
             if not (self.is_monotonic_increasing or self.is_monotonic_decreasing):
                 raise ValueError("index must be monotonic increasing or decreasing")
-            # error: "IndexEngine" has no attribute "get_indexer_with_fill"
-            engine = self._engine
-            with warnings.catch_warnings():
-                # TODO: We need to fix this. Casting to int64 in cython
-                warnings.filterwarnings("ignore", category=RuntimeWarning)
-                return engine.get_indexer_with_fill(  # type: ignore[union-attr]
-                    target=target._values,
-                    values=self._values,
-                    method=method,
-                    limit=limit,
-                )
+            encoded = self.append(target)._engine.values
+            self_encoded = Index(encoded[: len(self)])
+            target_encoded = Index(encoded[len(self) :])
+            return self_encoded._get_fill_indexer(
+                target_encoded, method, limit, tolerance
+            )
 
         if self.is_monotonic_increasing and target.is_monotonic_increasing:
             target_values = target._get_engine_target()