… pandas/_libs/index.pyi
… pandas/_libs/index.pyx