- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.2.0.rst` file if fixing a bug or adding a new feature.


```
from asv_bench.benchmarks.multiindex_object import Integer

b = Integer()
b.setup()

%timeit b.time_get_indexer_and_backfill()

# 238 ms ± 39.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)     -> main
# 50.6 ms ± 2.98 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  -> PR

%timeit b.time_get_indexer_and_pad()

# 191 ms ± 30.3 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)   -> main
# 50.4 ms ± 1.03 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  -> PR
```